Added separate functionality for part time staff
Also added a remainder function to allow for accurate additions of
multi-events within a day. For example if a hospital appt. is entered
and the user leaves at 13:00 for it they enter the event as a remainder
event duration. This assumes the start of the day is 0900 as there is no
working event when the hospital event is entered. When the working event
is added the hospital event is altered when a definitive start time is
entered.